feat: add CockroachDB compatibility to postgres driver
