--- cockroachdb/cockroachdb_test.go
@@ -1,537 +1,0 @@
-package cockroachdb
-
-import (
-	"context"
-	"fmt"
-	"net/url"
-	"os"
-	"strconv"
-	"testing"
-	"time"
-
-	"github.com/jackc/pgx/v5"
-	"github.com/stretchr/testify/require"
-	"github.com/testcontainers/testcontainers-go"
-	tccockroachdb "github.com/testcontainers/testcontainers-go/modules/cockroachdb"
-)
-
-const (
-	cockroachdbImage              = "cockroachdb/cockroach:latest-v23.2"
-	cockroachdbImageEnvVar string = "TEST_COCKROACHDB_IMAGE"
-	cockroachdbDatabase    string = "fiber"
-)
-
-func newTestConfig(t testing.TB) Config {
-	t.Helper()
-
-	ctx := context.Background()
-
-	img := cockroachdbImage
-	if imgFromEnv := os.Getenv(cockroachdbImageEnvVar); imgFromEnv != "" {
-		img = imgFromEnv
-	}
-
-	c, err := tccockroachdb.Run(ctx, img,
-		tccockroachdb.WithDatabase(cockroachdbDatabase),
-		tccockroachdb.WithInsecure(),
-	)
-	testcontainers.CleanupContainer(t, c)
-	require.NoError(t, err)
-
-	// ConnectionString() returns a stdlib registered key, not usable with pgxpool.
-	// Use ConnectionConfig() to get host/port/database and build a plain DSN.
-	connConfig, err := c.ConnectionConfig(ctx)
-	require.NoError(t, err)
-
-	dsn := fmt.Sprintf("postgresql://%s@%s:%d/%s?sslmode=disable",
-		connConfig.User, connConfig.Host, connConfig.Port, connConfig.Database)
-
-	return Config{
-		ConnectionURI: dsn,
-		Reset:         true,
-	}
-}
-
-func newTestStore(t testing.TB) *Storage {
-	t.Helper()
-
-	return newTestStoreWithConfig(t, newTestConfig(t))
-}
-
-func newTestStoreWithConfig(t testing.TB, cfg Config) *Storage {
-	t.Helper()
-
-	return New(cfg)
-}
-
-func TestNoCreateUser(t *testing.T) {
-	cfg := newTestConfig(t)
-
-	testStore := newTestStoreWithConfig(t, cfg)
-	defer testStore.Close()
-
-	ctx := context.Background()
-	conn := testStore.Conn()
-
-	// In insecure mode, passwords are not supported — create user without password.
-	username := "testuser" + strconv.Itoa(int(time.Now().UnixNano()))
-
-	_, err := conn.Exec(ctx, "CREATE USER "+username)
-	require.NoError(t, err)
-
-	_, err = conn.Exec(ctx, "GRANT CONNECT ON DATABASE "+cockroachdbDatabase+" TO "+username)
-	require.NoError(t, err)
-
-	_, err = conn.Exec(ctx, "GRANT USAGE ON SCHEMA public TO "+username)
-	require.NoError(t, err)
-
-	_, err = conn.Exec(ctx, "GRANT SELECT, INSERT, UPDATE, DELETE ON ALL TABLES IN SCHEMA public TO "+username)
-	require.NoError(t, err)
-
-	_, err = conn.Exec(ctx, "ALTER DEFAULT PRIVILEGES IN SCHEMA public GRANT SELECT, INSERT, UPDATE, DELETE ON TABLES TO "+username)
-	require.NoError(t, err)
-
-	u, parseErr := url.Parse(cfg.ConnectionURI)
-	require.NoError(t, parseErr)
-	u.User = url.User(username)
-	unprivilegedURI := u.String()
-
-	limitedCfg := Config{
-		ConnectionURI: unprivilegedURI,
-		Reset:         false,
-	}
-
-	limitedStore := New(limitedCfg)
-
-	defer func() {
-		limitedStore.Close()
-		_, _ = conn.Exec(ctx, "DROP USER "+username)
-	}()
-
-	t.Run("should set", func(t *testing.T) {
-		var (
-			key = "john" + strconv.Itoa(int(time.Now().UnixNano()))
-			val = []byte("doe" + strconv.Itoa(int(time.Now().UnixNano())))
-		)
-
-		err := limitedStore.Set(key, val, 0)
-		require.NoError(t, err)
-	})
-	t.Run("should set override", func(t *testing.T) {
-		var (
-			key = "john" + strconv.Itoa(int(time.Now().UnixNano()))
-			val = []byte("doe" + strconv.Itoa(int(time.Now().UnixNano())))
-		)
-		err := limitedStore.Set(key, val, 0)
-		require.NoError(t, err)
-		err = limitedStore.Set(key, val, 0)
-		require.NoError(t, err)
-	})
-	t.Run("should get", func(t *testing.T) {
-		var (
-			key = "john" + strconv.Itoa(int(time.Now().UnixNano()))
-			val = []byte("doe" + strconv.Itoa(int(time.Now().UnixNano())))
-		)
-		err := limitedStore.Set(key, val, 0)
-		require.NoError(t, err)
-		result, err := limitedStore.Get(key)
-		require.NoError(t, err)
-		require.Equal(t, val, result)
-	})
-	t.Run("should set expiration", func(t *testing.T) {
-		var (
-			key = "john" + strconv.Itoa(int(time.Now().UnixNano()))
-			val = []byte("doe" + strconv.Itoa(int(time.Now().UnixNano())))
-			exp = 100 * time.Millisecond
-		)
-		err := limitedStore.Set(key, val, exp)
-		require.NoError(t, err)
-	})
-	t.Run("should get expired", func(t *testing.T) {
-		var (
-			key = "john" + strconv.Itoa(int(time.Now().UnixNano()))
-			val = []byte("doe" + strconv.Itoa(int(time.Now().UnixNano())))
-			exp = 100 * time.Millisecond
-		)
-		err := limitedStore.Set(key, val, exp)
-		require.NoError(t, err)
-		time.Sleep(200 * time.Millisecond)
-		result, err := limitedStore.Get(key)
-		require.NoError(t, err)
-		require.Zero(t, len(result))
-	})
-	t.Run("should get not exists", func(t *testing.T) {
-		result, err := limitedStore.Get("nonexistentkey")
-		require.NoError(t, err)
-		require.Zero(t, len(result))
-	})
-	t.Run("should delete", func(t *testing.T) {
-		var (
-			key = "john" + strconv.Itoa(int(time.Now().UnixNano()))
-			val = []byte("doe" + strconv.Itoa(int(time.Now().UnixNano())))
-		)
-		err := limitedStore.Set(key, val, 0)
-		require.NoError(t, err)
-		err = limitedStore.Delete(key)
-		require.NoError(t, err)
-		result, err := limitedStore.Get(key)
-		require.NoError(t, err)
-		require.Zero(t, len(result))
-	})
-}
-
-func Test_Should_Panic_On_Wrong_Schema(t *testing.T) {
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	_, err := testStore.Conn().Exec(context.Background(), `
-			CREATE TABLE IF NOT EXISTS test_schema_table (
-				k  VARCHAR(64) PRIMARY KEY NOT NULL DEFAULT '',
-				v  BYTEA NOT NULL,
-				e  VARCHAR(64) NOT NULL DEFAULT ''
-			);
-		`)
-	require.NoError(t, err)
-	defer func() {
-		_, err := testStore.Conn().Exec(context.Background(), "DROP TABLE IF EXISTS test_schema_table;")
-		require.NoError(t, err)
-	}()
-
-	require.Panics(t, func() {
-		testStore.checkSchema("test_schema_table")
-	})
-}
-
-func Test_CockroachDB_Set(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, 0)
-	require.NoError(t, err)
-}
-
-func Test_CockroachDB_SetWithContext(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.SetWithContext(ctx, key, val, 0)
-	require.ErrorIs(t, err, context.Canceled)
-}
-
-func Test_CockroachDB_Set_Override(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, 0)
-	require.NoError(t, err)
-
-	err = testStore.Set(key, val, 0)
-	require.NoError(t, err)
-}
-
-func Test_CockroachDB_Get(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, 0)
-	require.NoError(t, err)
-
-	result, err := testStore.Get(key)
-	require.NoError(t, err)
-	require.Equal(t, val, result)
-}
-
-func Test_CockroachDB_GetWithContext(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, 0)
-	require.NoError(t, err)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	result, err := testStore.GetWithContext(ctx, key)
-	require.ErrorIs(t, err, context.Canceled)
-	require.Zero(t, len(result))
-}
-
-func Test_CockroachDB_Set_Expiration(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-		exp = 1 * time.Second
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, exp)
-	require.NoError(t, err)
-
-	time.Sleep(1100 * time.Millisecond)
-
-	result, err := testStore.Get(key)
-	require.NoError(t, err)
-	require.Zero(t, len(result), "Key should have expired")
-}
-
-func Test_CockroachDB_Get_Expired(t *testing.T) {
-	key := "john"
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	result, err := testStore.Get(key)
-	require.NoError(t, err)
-	require.Zero(t, len(result))
-}
-
-func Test_CockroachDB_Get_NotExist(t *testing.T) {
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	result, err := testStore.Get("notexist")
-	require.NoError(t, err)
-	require.Zero(t, len(result))
-}
-
-func Test_CockroachDB_Delete(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, 0)
-	require.NoError(t, err)
-
-	err = testStore.Delete(key)
-	require.NoError(t, err)
-
-	result, err := testStore.Get(key)
-	require.NoError(t, err)
-	require.Zero(t, len(result))
-}
-
-func Test_CockroachDB_DeleteWithContext(t *testing.T) {
-	var (
-		key = "john"
-		val = []byte("doe")
-	)
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set(key, val, 0)
-	require.NoError(t, err)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	err = testStore.DeleteWithContext(ctx, key)
-	require.ErrorIs(t, err, context.Canceled)
-}
-
-func Test_CockroachDB_Reset(t *testing.T) {
-	val := []byte("doe")
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set("john1", val, 0)
-	require.NoError(t, err)
-
-	err = testStore.Set("john2", val, 0)
-	require.NoError(t, err)
-
-	err = testStore.Reset()
-	require.NoError(t, err)
-
-	result, err := testStore.Get("john1")
-	require.NoError(t, err)
-	require.Zero(t, len(result))
-
-	result, err = testStore.Get("john2")
-	require.NoError(t, err)
-	require.Zero(t, len(result))
-}
-
-func Test_CockroachDB_ResetWithContext(t *testing.T) {
-	val := []byte("doe")
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set("john1", val, 0)
-	require.NoError(t, err)
-
-	err = testStore.Set("john2", val, 0)
-	require.NoError(t, err)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	err = testStore.ResetWithContext(ctx)
-	require.ErrorIs(t, err, context.Canceled)
-
-	result, err := testStore.Get("john1")
-	require.NoError(t, err)
-	require.NotZero(t, len(result))
-
-	result, err = testStore.Get("john2")
-	require.NoError(t, err)
-	require.NotZero(t, len(result))
-}
-
-func Test_CockroachDB_GC(t *testing.T) {
-	testVal := []byte("doe")
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set("john", testVal, time.Nanosecond)
-	require.NoError(t, err)
-
-	testStore.gc(time.Now())
-	row := testStore.db.QueryRow(context.Background(), testStore.sqlSelect, "john")
-	err = row.Scan(nil, nil)
-	require.Equal(t, pgx.ErrNoRows, err)
-
-	err = testStore.Set("john", testVal, 0)
-	require.NoError(t, err)
-
-	testStore.gc(time.Now())
-	val, err := testStore.Get("john")
-	require.NoError(t, err)
-	require.Equal(t, testVal, val)
-}
-
-func Test_CockroachDB_Non_UTF8(t *testing.T) {
-	val := []byte("0xF5")
-
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	err := testStore.Set("0xF6", val, 0)
-	require.NoError(t, err)
-
-	result, err := testStore.Get("0xF6")
-	require.NoError(t, err)
-	require.Equal(t, val, result)
-}
-
-func Test_CockroachDB_Panic_On_No_Server(t *testing.T) {
-	require.Panics(t, func() {
-		_ = New(Config{
-			Reset: true,
-		})
-	}, "Expected panic when connecting to CockroachDB with no server running")
-}
-
-func Test_CockroachDB_Conn(t *testing.T) {
-	testStore := newTestStore(t)
-	defer testStore.Close()
-
-	require.True(t, testStore.Conn() != nil)
-}
-
-func Test_CockroachDB_Close(t *testing.T) {
-	testStore := newTestStore(t)
-	require.NoError(t, testStore.Close())
-}
-
-func Benchmark_CockroachDB_Set(b *testing.B) {
-	testStore := newTestStore(b)
-	defer testStore.Close()
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	var err error
-	for i := 0; i < b.N; i++ {
-		err = testStore.Set("john", []byte("doe"), 0)
-	}
-
-	require.NoError(b, err)
-}
-
-func Benchmark_CockroachDB_Get(b *testing.B) {
-	testStore := newTestStore(b)
-	defer testStore.Close()
-
-	err := testStore.Set("john", []byte("doe"), 0)
-	require.NoError(b, err)
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		_, err = testStore.Get("john")
-	}
-
-	require.NoError(b, err)
-}
-
-func Benchmark_CockroachDB_SetAndDelete(b *testing.B) {
-	testStore := newTestStore(b)
-	defer testStore.Close()
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	var err error
-	for i := 0; i < b.N; i++ {
-		_ = testStore.Set("john", []byte("doe"), 0)
-		err = testStore.Delete("john")
-	}
-
-	require.NoError(b, err)
-}
-
-func Test_CockroachDB_CustomTable_WithSchema(t *testing.T) {
-	cfg := newTestConfig(t)
-	cfg.Table = "public.fiber_storage_" + strconv.Itoa(int(time.Now().UnixNano()))
-
-	store := newTestStoreWithConfig(t, cfg)
-	defer store.Close()
-
-	val := []byte("doe")
-	err := store.Set("john", val, 0)
-	require.NoError(t, err)
-
-	result, err := store.Get("john")
-	require.NoError(t, err)
-	require.Equal(t, val, result)
-}
-
